+++ src/test/multisig_tests.cpp
@@ -4,7 +4,7 @@
 
 #include "key.h"
 #include "keystore.h"
-#include "main.h"
+#include "policy/policy.h"
 #include "script/script.h"
 #include "script/script_error.h"
 #include "script/interpreter.h"
